chore(1892): Moved loader to a separate component
