+++ src/components/bouncing-loader/index.ts
@@ -1,0 +1,1 @@
+export {default} from './bouncing-loader';
